refactor: trim __init__.py to Tier 1 API; split presets from factories (#185)
Move the 10 preset geometry factory functions from factories.py to a new
presets.py module.  Trim __init__.py from 97 exported names down to 23
(Tier 1 — routine session essentials).  All other names remain accessible
via their submodule (ahd.display.fmt(), ahd.radiation.energy_to_wavelength(),
ahd.presets.fourcv(), etc.).

Export register_geometry as a Tier 1 name, resolving #145.

Update all test files, documentation code blocks, Sphinx cross-references,
pyproject.toml entry points, and AGENTS.md to reflect the new structure.

Contributed by: OpenCode (argo/claudeopus46)
