+++ docs/source/geometries/kappa4ch.md
@@ -12,14 +12,14 @@ Four-circle kappa diffractometer, horizontal scattering plane. Kappa axis tilted
 ```python
 import ad_hoc_diffractometer as ahd
 
-g = ahd.kappa4ch()
+g = ahd.presets.kappa4ch()
 g.wavelength = 1.0  # Å
 print(g.summary())
 ```
 
 ## Pre-built geometry definition
 
-This geometry is defined by the {func}`~ad_hoc_diffractometer.factories.kappa4ch` factory
+This geometry is defined by the {func}`~ad_hoc_diffractometer.presets.kappa4ch` factory
 function — see the [source](https://github.com/prjemian/ad_hoc_diffractometer/blob/main/src/ad_hoc_diffractometer/factories.py#L964) for the complete stage
 and mode configuration.
 
@@ -116,7 +116,7 @@ the stored target.  See {doc}`../howto/surface`.
 
 ## API reference
 
-- {func}`~ad_hoc_diffractometer.factories.kappa4ch`
+- {func}`~ad_hoc_diffractometer.presets.kappa4ch`
 - {class}`~ad_hoc_diffractometer.geometry.AdHocDiffractometer`
 - {class}`~ad_hoc_diffractometer.mode.ConstraintSet`
 - {class}`~ad_hoc_diffractometer.mode.BisectConstraint`
