+++ docs/source/howto/modes.md
@@ -12,7 +12,7 @@ See [Concepts](../concepts.md) for background, and
 ```python
 import ad_hoc_diffractometer as ahd
 
-g = ahd.fourcv()
+g = ahd.presets.fourcv()
 print(list(g.modes.keys()))
 # ['bisecting', 'fixed_chi', 'fixed_phi', 'constant_omega',
 #  'psi_constant', 'double_diffraction']
@@ -91,7 +91,7 @@ solutions = g.forward(1, 0, 0)
 ## Inspect a mode's constraints
 
 ```python
-g = ahd.fourcv()
+g = ahd.presets.fourcv()
 cs = g.modes["fixed_chi"]
 
 print(cs)
