+++ pyproject.toml
@@ -17,16 +17,16 @@ dependencies = [
 ]
 
 [project.entry-points."ad_hoc_diffractometer.geometries"]
-fivec    = "ad_hoc_diffractometer.factories:fivec"
-fourch   = "ad_hoc_diffractometer.factories:fourch"
-fourcv   = "ad_hoc_diffractometer.factories:fourcv"
-kappa4ch = "ad_hoc_diffractometer.factories:kappa4ch"
-kappa4cv = "ad_hoc_diffractometer.factories:kappa4cv"
-kappa6c  = "ad_hoc_diffractometer.factories:kappa6c"
-psic     = "ad_hoc_diffractometer.factories:psic"
-s2d2     = "ad_hoc_diffractometer.factories:s2d2"
-sixc     = "ad_hoc_diffractometer.factories:sixc"
-zaxis    = "ad_hoc_diffractometer.factories:zaxis"
+fivec    = "ad_hoc_diffractometer.presets:fivec"
+fourch   = "ad_hoc_diffractometer.presets:fourch"
+fourcv   = "ad_hoc_diffractometer.presets:fourcv"
+kappa4ch = "ad_hoc_diffractometer.presets:kappa4ch"
+kappa4cv = "ad_hoc_diffractometer.presets:kappa4cv"
+kappa6c  = "ad_hoc_diffractometer.presets:kappa6c"
+psic     = "ad_hoc_diffractometer.presets:psic"
+s2d2     = "ad_hoc_diffractometer.presets:s2d2"
+sixc     = "ad_hoc_diffractometer.presets:sixc"
+zaxis    = "ad_hoc_diffractometer.presets:zaxis"
 
 [project.optional-dependencies]
 dev = [
