+++ tests/conftest.py
@@ -15,8 +15,8 @@ helpers.py — import those explicitly with `from helpers import ...`.
 
 import pytest
 
-from ad_hoc_diffractometer import psic
-from ad_hoc_diffractometer import set_precision
+from ad_hoc_diffractometer.display import set_precision
+from ad_hoc_diffractometer.presets import psic
 
 _DISPLAY_DEFAULT = 6  # matches display._DEFAULT_PRECISION
