+++ tests/test_rotation.py
@@ -15,10 +15,10 @@ from helpers import Rx
 from helpers import Ry
 from helpers import Rz
 
-from ad_hoc_diffractometer import XHAT
-from ad_hoc_diffractometer import YHAT
-from ad_hoc_diffractometer import ZHAT
-from ad_hoc_diffractometer import rotation_matrix
+from ad_hoc_diffractometer.constants import XHAT
+from ad_hoc_diffractometer.constants import YHAT
+from ad_hoc_diffractometer.constants import ZHAT
+from ad_hoc_diffractometer.rotation import rotation_matrix
 
 # ---------------------------------------------------------------------------
 # rotation_matrix() — correctness
