+++ tests/test_stage.py
@@ -21,12 +21,12 @@ from helpers import Rx
 from helpers import Ry
 from helpers import Rz
 
-from ad_hoc_diffractometer import XHAT
-from ad_hoc_diffractometer import YHAT
-from ad_hoc_diffractometer import ZHAT
 from ad_hoc_diffractometer import AdHocDiffractometer
-from ad_hoc_diffractometer import Stage
-from ad_hoc_diffractometer import fourcv
+from ad_hoc_diffractometer.constants import XHAT
+from ad_hoc_diffractometer.constants import YHAT
+from ad_hoc_diffractometer.constants import ZHAT
+from ad_hoc_diffractometer.presets import fourcv
+from ad_hoc_diffractometer.stage import Stage
 
 # ---------------------------------------------------------------------------
 # Stage construction
